FlaskAPI/requirements.txt
-i https://pypi.org/simple

certifi==2020.11.8

cffi==1.14.4

chardet==3.0.4

click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

cryptography==3.2.1

ecdsa==0.14.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

flask==1.1.2

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

msal==1.6.0

pyasn1==0.4.8

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyjwt[crypto]==1.7.1

python-dotenv==0.15.0

python-jose==3.2.0

requests==2.25.0

rsa==4.6; python_version >= '3.5' and python_version < '4'

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

